Move query classes to their own package
